fixed bugs, created stream class
