+++ httpy/stream.py
@@ -1,0 +1,28 @@
+class Stream:
+    def __init__(self, generator):
+        self.gen = generator
+        self.buffer = bytearray()
+
+    def read(self, nbytes):
+        while len(self.buffer) < nbytes:
+            try:
+                self.buffer.extend(next(self.gen))
+            except StopIteration:
+                break
+        self.buffer, result = self.buffer[:nbytes], self.buffer[nbytes:]
+        return bytes(result)
+
+
+class AsyncStream:
+    def __init__(self, generator):
+        self.gen = generator
+        self.buffer = bytearray()
+
+    async def read(self, nbytes):
+        while len(self.buffer) < nbytes:
+            try:
+                self.buffer.extend(await anext(self.gen))
+            except StopIteration:
+                break
+        self.buffer, result = self.buffer[:nbytes], self.buffer[nbytes:]
+        return bytes(result)
